% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osn_utils.R
\name{available_mcm}
\alias{available_mcm}
\title{list available mcmicro spatial proteomics for GBCC}
\usage{
available_mcm(local = FALSE)
}
\arguments{
\item{local}{logical(1), if TRUE, assume available at Sys.getenv("MCM_LOCAL")}
}
\value{
vector of paths
}
\description{
list available mcmicro spatial proteomics for GBCC
}
