% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overviewPlots.R
\name{BCoverviewPlots}
\alias{BCoverviewPlots}
\title{provide an overview plot of all breast cancer samples and cell type distributions}
\usage{
BCoverviewPlots(BPPARAM = BiocParallel::MulticoreParam(4))
}
\description{
provide an overview plot of all breast cancer samples and cell type distributions
}
\note{
Returns a ggplot object
}
\examples{
BCoverviewPlots()
}
