% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process1.R
\name{browseOne}
\alias{browseOne}
\title{define shiny app to explore one sample}
\usage{
browseOne(input_processed, targets = targets_june())
}
\arguments{
\item{input_processed}{output of `process_mcmicro`}

\item{targets}{character() vector}
}
\description{
define shiny app to explore one sample
}
\examples{
pa = system.file("h5ad", "Galaxy37-LSP12022-A3.h5ad", package="GBCC.mcm")
a3 = process_mcmicro(pa)
# be sure to run example(..., ask=FALSE)
if (interactive()) {
  browseOne(a3)
}
}
