% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_mcm_path}
\alias{get_mcm_path}
\title{get path to mcmicro h5ad in package}
\usage{
get_mcm_path(tag = "A4")
}
\arguments{
\item{tag}{character(1) one of A3, A4, B3, B4}
}
\description{
get path to mcmicro h5ad in package
}
\note{
A3 = Triple Negative Breast Cancer; A4 = Luminal B HER2+ Breast Cancer; 
B3 = Luminal B HER- Breast Cancer; B4 = normal jejunum
}
