% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mcmicro.R
\name{process_mcmicro}
\alias{process_mcmicro}
\title{Use Voyager spatial statistics support with MCMICRO proteomics experiments}
\usage{
process_mcmicro(
  h5ad,
  coordnames = c("X_centroid", "Y_centroid"),
  assay4quants = "X",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{h5ad}{character(1) path to an MCMICRO output in h5ad format}

\item{coordnames}{character() names in colData of transformed h5ad used to specify X, Y}

\item{assay4quants}{character(1) name in assayNames of transformed h5ad used as protein quantification}

\item{verbose}{logical(1) will message as stages complete if TRUE}
}
\description{
Use Voyager spatial statistics support with MCMICRO proteomics experiments
}
\examples{
pa = get_mcm_path("m53.1")
mcm53.1 = process_mcmicro(pa)
features_use = c("NCAM", "FOXP3", "CD8A", "LDH")
Voyager::plotLocalResult(mcm53.1, "localG", features = features_use,
                colGeometryName = "centroids", divergent = TRUE,
                diverge_center = 0)
}
