% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overviewPlots.R
\name{processBC}
\alias{processBC}
\title{process Breast Cancer data from MCMICRO}
\usage{
processBC(BPPARAM = BiocParallel::MulticoreParam(4))
}
\arguments{
\item{BPPARAM}{a BiocParallel *Param object}
}
\value{
a list of SpatialFeatureExperiments
}
\description{
process Breast Cancer data from MCMICRO
}
