% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mcmicro.R
\name{default_uruns}
\alias{default_uruns}
\title{create a default configuration for several runUnivariate tasks}
\usage{
default_uruns(
  assay4quants = "X",
  colGraphName = "spatNeigh",
  BPPARAM = BiocParallel::bpparam()
)
}
\description{
create a default configuration for several runUnivariate tasks
}
\note{
dots arguments must be present for all components
}
