% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_mcm_path}
\alias{get_mcm_path}
\title{get path to mcmicro h5ad in package}
\usage{
get_mcm_path(tag = "m62.1")
}
\arguments{
\item{tag}{character(1) one of m62.1, ..., m62.4, m62.1}
}
\description{
get path to mcmicro h5ad in package
}
