% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osn_utils.R
\name{path_to_zipped_mcm}
\alias{path_to_zipped_mcm}
\title{retrieve a specified zipped experiment from cache, downloading to cache if necessary}
\usage{
path_to_zipped_mcm(
  experiment = "m53.1.sfe.zip",
  local = FALSE,
  cache = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{experiment}{character(1)}
}
\description{
retrieve a specified zipped experiment from cache, downloading to cache if necessary
}
